refactor(archive-preview): extract internal and external archive listing
