+++ src/preview/container/archive/external.rs
@@ -1,0 +1,127 @@
+use super::common::{normalize_archive_path, parse_key_value_line, parse_u64};
+use super::*;
+use crate::preview::ArchiveFormat;
+use std::{collections::BTreeMap, path::Path, process::Command};
+
+pub(super) fn fallback_single_file_archive_entry(
+    path: &Path,
+    format: ArchiveFormat,
+) -> Option<ArchiveEntry> {
+    if !matches!(
+        format,
+        ArchiveFormat::Gzip | ArchiveFormat::Xz | ArchiveFormat::Bzip2 | ArchiveFormat::Zstd
+    ) {
+        return None;
+    }
+
+    let name = path.file_stem()?.to_str()?;
+    let path = normalize_archive_path(name, false)?;
+    Some(ArchiveEntry {
+        path,
+        is_dir: false,
+    })
+}
+
+pub(super) fn collect_archive_entries_with_bsdtar(path: &Path) -> Option<Vec<ArchiveEntry>> {
+    let output = Command::new("bsdtar").arg("-tf").arg(path).output().ok()?;
+    if !output.status.success() {
+        return None;
+    }
+    Some(normalize_archive_entries(
+        String::from_utf8_lossy(&output.stdout).lines(),
+        false,
+    ))
+}
+
+pub(super) fn collect_archive_listing_with_7z(
+    path: &Path,
+) -> Option<(ArchiveMetadata, Vec<ArchiveEntry>)> {
+    let output = Command::new("7z")
+        .arg("l")
+        .arg("-slt")
+        .arg(path)
+        .output()
+        .ok()?;
+    if !output.status.success() {
+        return None;
+    }
+    parse_7z_listing(&String::from_utf8_lossy(&output.stdout))
+}
+
+pub(super) fn parse_7z_listing(output: &str) -> Option<(ArchiveMetadata, Vec<ArchiveEntry>)> {
+    let mut metadata = ArchiveMetadata::default();
+    let mut entries = Vec::new();
+    let mut in_entries = false;
+    let mut current = BTreeMap::<String, String>::new();
+
+    for raw_line in output.lines() {
+        let line = raw_line.trim_end();
+        if line == "----------" {
+            in_entries = true;
+            continue;
+        }
+
+        if !in_entries {
+            if let Some((key, value)) = parse_key_value_line(line) {
+                match key {
+                    "Type" => metadata.format_label = Some(value.to_string()),
+                    "Physical Size" => metadata.physical_size = parse_u64(value),
+                    "Comment" if !value.is_empty() => metadata.comment = Some(value.to_string()),
+                    _ => {}
+                }
+            }
+            continue;
+        }
+
+        if line.is_empty() {
+            push_7z_entry(&mut current, &mut entries, &mut metadata);
+            continue;
+        }
+
+        if let Some((key, value)) = parse_key_value_line(line) {
+            current.insert(key.to_string(), value.to_string());
+        }
+    }
+    push_7z_entry(&mut current, &mut entries, &mut metadata);
+
+    if entries.is_empty()
+        && metadata.format_label.is_none()
+        && metadata.physical_size.is_none()
+        && metadata.comment.is_none()
+    {
+        None
+    } else {
+        Some((metadata, entries))
+    }
+}
+
+fn push_7z_entry(
+    current: &mut BTreeMap<String, String>,
+    entries: &mut Vec<ArchiveEntry>,
+    metadata: &mut ArchiveMetadata,
+) {
+    if current.is_empty() {
+        return;
+    }
+
+    let path = current.get("Path").cloned();
+    let is_dir = current.get("Folder").is_some_and(|value| value == "+")
+        || current
+            .get("Attributes")
+            .is_some_and(|value| value.starts_with('D'));
+
+    if let Some(path) = path.and_then(|path| normalize_archive_path(&path, false)) {
+        entries.push(ArchiveEntry { path, is_dir });
+    }
+
+    if let Some(size) = current.get("Size").and_then(|value| parse_u64(value)) {
+        metadata.unpacked_size = Some(metadata.unpacked_size.unwrap_or(0).saturating_add(size));
+    }
+    if let Some(size) = current
+        .get("Packed Size")
+        .and_then(|value| parse_u64(value))
+    {
+        metadata.compressed_size = Some(metadata.compressed_size.unwrap_or(0).saturating_add(size));
+    }
+    current.clear();
+}
